+++ src/java/org/malbino/orion/servlets/ListaInscritosCarrera.java
@@ -39,8 +39,8 @@ import org.malbino.orion.facades.InscritoFacade;
  *
  * @author tincho
  */
-@WebServlet(name = "LibroInscripciones", urlPatterns = {"/reportes/LibroInscripciones"})
-public class LibroInscripciones extends HttpServlet {
+@WebServlet(name = "ListaInscritosCarrera", urlPatterns = {"/reportes/ListaInscritosCarrera"})
+public class ListaInscritosCarrera extends HttpServlet {
 
     private static final String CONTENIDO_PDF = "application/pdf";
 
@@ -91,7 +91,7 @@ public class LibroInscripciones extends HttpServlet {
 
                 document.close();
             } catch (IOException | DocumentException ex) {
-                Logger.getLogger(LibroInscripciones.class.getName()).log(Level.SEVERE, null, ex);
+                Logger.getLogger(ListaInscritosCarrera.class.getName()).log(Level.SEVERE, null, ex);
             }
 
         }
@@ -107,25 +107,19 @@ public class LibroInscripciones extends HttpServlet {
         image.setAlignment(Image.ALIGN_CENTER);
         PdfPCell cell = new PdfPCell();
         cell.addElement(image);
-        cell.setRowspan(4);
+        cell.setRowspan(5);
         cell.setColspan(20);
         cell.setVerticalAlignment(PdfPCell.ALIGN_MIDDLE);
         cell.setBorder(Rectangle.NO_BORDER);
         table.addCell(cell);
 
-        cell = new PdfPCell(new Phrase("LIBRO DE INSCRIPCIONES,", TITULO));
+        cell = new PdfPCell(new Phrase("LISTA DE INSCRITOS POR CARRERA,", TITULO));
         cell.setHorizontalAlignment(PdfPCell.ALIGN_LEFT);
         cell.setColspan(80);
         cell.setBorder(Rectangle.NO_BORDER);
         cell.setBackgroundColor(BaseColor.LIGHT_GRAY);
         table.addCell(cell);
-        
-        cell = new PdfPCell(new Phrase(carrera.getCampus().getInstituto().getNombreRegulador(), SUBTITULO));
-        cell.setHorizontalAlignment(PdfPCell.ALIGN_LEFT);
-        cell.setColspan(80);
-        cell.setBorder(Rectangle.NO_BORDER);
-        table.addCell(cell);
-        
+
         cell = new PdfPCell(new Phrase(gestionAcademica.toString(), SUBTITULO));
         cell.setHorizontalAlignment(PdfPCell.ALIGN_LEFT);
         cell.setColspan(80);
@@ -168,25 +162,25 @@ public class LibroInscripciones extends HttpServlet {
         cell.setBackgroundColor(BaseColor.LIGHT_GRAY);
         table.addCell(cell);
 
-        cell = new PdfPCell(new Phrase("Tipo", NEGRITA));
+        cell = new PdfPCell(new Phrase("Telfono", NEGRITA));
         cell.setHorizontalAlignment(PdfPCell.ALIGN_CENTER);
         cell.setColspan(10);
         cell.setBackgroundColor(BaseColor.LIGHT_GRAY);
         table.addCell(cell);
 
-        cell = new PdfPCell(new Phrase("Fecha", NEGRITA));
+        cell = new PdfPCell(new Phrase("Celular", NEGRITA));
         cell.setHorizontalAlignment(PdfPCell.ALIGN_CENTER);
         cell.setColspan(10);
         cell.setBackgroundColor(BaseColor.LIGHT_GRAY);
         table.addCell(cell);
 
-        cell = new PdfPCell(new Phrase("Firma", NEGRITA));
+        cell = new PdfPCell(new Phrase("Email", NEGRITA));
         cell.setHorizontalAlignment(PdfPCell.ALIGN_CENTER);
         cell.setColspan(30);
         cell.setBackgroundColor(BaseColor.LIGHT_GRAY);
         table.addCell(cell);
 
-        List<Inscrito> listaInscritos = inscritoFacade.listaInscritos(gestionAcademica.getId_gestionacademica(), carrera.getId_carrera());
+        List<Inscrito> listaInscritos = inscritoFacade.listaInscritosCarrera(gestionAcademica.getId_gestionacademica(), carrera.getId_carrera());
         for (int i = 0; i < listaInscritos.size(); i++) {
             Inscrito inscrito = listaInscritos.get(i);
 
@@ -208,23 +202,34 @@ public class LibroInscripciones extends HttpServlet {
             cell.setColspan(35);
             table.addCell(cell);
 
-            cell = new PdfPCell(new Phrase(inscrito.getTipo().toString(), NORMAL));
+            if (inscrito.getEstudiante().getTelefono() != null) {
+                cell = new PdfPCell(new Phrase(String.valueOf(inscrito.getEstudiante().getTelefono()), NORMAL));
+            } else {
+                cell = new PdfPCell(new Phrase(" ", NORMAL));
+            }
             cell.setHorizontalAlignment(PdfPCell.ALIGN_CENTER);
             cell.setVerticalAlignment(PdfPCell.ALIGN_MIDDLE);
             cell.setColspan(10);
             table.addCell(cell);
 
-            cell = new PdfPCell(new Phrase(inscrito.fecha_ddMMyyyy(), NORMAL));
+            if (inscrito.getEstudiante().getCelular() != null) {
+                cell = new PdfPCell(new Phrase(String.valueOf(inscrito.getEstudiante().getCelular()), NORMAL));
+            } else {
+                cell = new PdfPCell(new Phrase(" ", NORMAL));
+            }
             cell.setHorizontalAlignment(PdfPCell.ALIGN_CENTER);
             cell.setVerticalAlignment(PdfPCell.ALIGN_MIDDLE);
             cell.setColspan(10);
             table.addCell(cell);
 
-            cell = new PdfPCell(new Phrase(" ", NORMAL));
-            cell.setHorizontalAlignment(PdfPCell.ALIGN_CENTER);
+            if (inscrito.getEstudiante().getEmail() != null) {
+                cell = new PdfPCell(new Phrase(inscrito.getEstudiante().getEmail(), NORMAL));
+            } else {
+                cell = new PdfPCell(new Phrase(" ", NORMAL));
+            }
+            cell.setHorizontalAlignment(PdfPCell.ALIGN_LEFT);
             cell.setVerticalAlignment(PdfPCell.ALIGN_MIDDLE);
             cell.setColspan(30);
-            cell.setFixedHeight(30);
             table.addCell(cell);
         }