+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0] - 2021-03-03
+### Added
+- Se añadio el Reporte de Inscritos por Carrra
+- Se añadio el Reporte de Inscritos por Paralelo
+- Se añadio el Reporte de Inscritos por Grupo
+### Changed
+### Removed
+### Fixed
 
 ## [1.0.0] - 2021-02-21
 ### Added